Moved search sub components and updated imports
