+++ src/components/search/search-articles/AtpSearchArticles.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/main.scss';
+@import '../../../assets/styles/main.scss';
 
 .atp-search-articles {
   &__title {
