+++ src/components/search/search-articles/AtpSearchArticles.tsx
@@ -1,8 +1,8 @@
 import React, { FC } from 'react';
-import { strings } from '../../constants/strings';
-import { Article } from '../../models/article.model';
-import AtpArticleCard from '../article-card/AtpArticleCard';
-import AtpText from '../text/AtpText';
+import { strings } from '../../../constants/strings';
+import { Article } from '../../../models/article.model';
+import AtpArticleCard from '../../article-card/AtpArticleCard';
+import AtpText from '../../text/AtpText';
 import './AtpSearchArticles.scss';
 
 type Props = {
