+++ src/components/search/search-box/AtpSearchBox.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/main.scss';
+@import '../../../assets/styles/main.scss';
 
 .atp-search-box {
   position: fixed;
