+++ src/components/search/search-box/AtpSearchBox.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import './AtpSearchBox.scss';
-import { strings } from '../../constants/strings';
-import AtpButton from '../button/AtpButton';
-import AtpLink from '../link-atp/AtpLink';
+import { strings } from '../../../constants/strings';
+import AtpButton from '../../button/AtpButton';
+import AtpLink from '../../link-atp/AtpLink';
 import { IoSearchOutline } from 'react-icons/io5';
 import { useNavigate } from 'react-router-dom';
