+++ src/components/search/search-brands/AtpSearchBrands.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/main.scss';
+@import '../../../assets/styles/main.scss';
 
 .atp-search-brands {
   text-align: left;
