+++ src/components/search/search-categories/AtpSearchCategories.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/main.scss';
+@import '../../../assets/styles/main.scss';
 
 .atp-search-categories {
   text-align: left;
