+++ src/components/search/search-colors/AtpSearchColors.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/main.scss';
+@import '../../../assets/styles/main.scss';
 
 .atp-search-colors {
   text-align: left;
