+++ src/components/search/search-selection/AtpSearchFilters.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/main.scss';
+@import '../../../assets/styles/main.scss';
 
 .atp-search-filters-mobile {
   position: fixed;
