+++ src/components/search/search-sort/AtpSearchSort.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/main.scss';
+@import '../../../assets/styles/main.scss';
 
 .atp-search-sort-mobile {
   position: fixed;
