+++ src/components/search/search-tabs/AtpSearchTabs.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/main.scss';
+@import '../../../assets/styles/main.scss';
 
 .atp-search-tabs {
   list-style: none;
